Updated Ratings to single point decimals

+++ Shufl.Domain.Entities/GroupSuggestionRating.cs
@@ -9,11 +9,11 @@ namespace Shufl.Domain.Entities
     {
         public Guid Id { get; set; }
         public Guid GroupSuggestionId { get; set; }
-        public byte OverallRating { get; set; }
-        public byte LyricsRating { get; set; }
-        public byte VocalsRating { get; set; }
-        public byte InstrumentalsRating { get; set; }
-        public byte CompositionRating { get; set; }
+        public decimal OverallRating { get; set; }
+        public decimal LyricsRating { get; set; }
+        public decimal VocalsRating { get; set; }
+        public decimal InstrumentalsRating { get; set; }
+        public decimal CompositionRating { get; set; }
         public string Comment { get; set; }
         public DateTime CreatedOn { get; set; }
         public Guid CreatedBy { get; set; }
